Core/DataStores: Updated db2 structures to 8.2.5
Port From (https://github.com/TrinityCore/TrinityCore/commit/3ac790287aba5f7d7c3bccf79e608de9119e461a)

+++ Source/Game/DataStorage/Structs/U_Records.cs
@@ -79,7 +79,7 @@ namespace Game.DataStorage
         public string ToolTip;
         public uint MinPower;
         public uint MaxPower;
-        public ushort StartPower;
+        public uint StartPower;
         public byte CenterPower;
         public float RegenerationPeace;
         public float RegenerationCombat;
